Modularize. Isolate start war command.
